feat: remove openthread border router as it can't currently run on K8s
It's currently not possible to have it run correctly with the level of
access to network hardware it needs. Works just fine on plain docker,
but crashes (or can't start) on kubernetes, even as root, with
hostNetwork, all permissions, etc. Maybe in the future.
Possible workaround is either a Pi or an LXC/VM on Proxmox